Fix Python 3.12 test failures and refactor codebase. Break apart some monoliths.
Fix pylint issues

Expand .gitignore
